web: 可写工作区 + 块视图，前端文件管理与 IndexedDB 持久化
- workspace.rs：可写工作区（默认 examples，与 UI/ 同发现序），保存/删除/新建/改名/建目录；
  相对路径逐段校验 + canonicalize 根内断言 + 4MB 上限
- astdoc.rs：ast/schema 信封（AST 结构化导出 + 原语/库 schema），span 全量携带、原文保真，
  作为块视图 IR；解析失败前端降级只读
- ws.rs/mod.rs：新增 workspace/tree/browse/rename/mkdir 信封
- 前端：blocks.tsx（块视图）、files.tsx（文件管理）、store.ts（IndexedDB 持久化）、markdown.ts；
  App.tsx 重构 + panels.tsx/cm.ts 增强
- packet-dsl：lib_functions 枚举全部函数（含未导出）供 LSP 补全；lsp.rs 补全增强
- 图标：icon.ico/icon.svg + winresource 嵌入 exe；justfile Windows 本机/交叉自动分派

+++ crates/prping-core/src/ping/trace/dns.rs
@@ -104,7 +104,8 @@ fn reverse_dns(ip: IpAddr) -> Option<String> {
             1024,
             serv.as_mut_ptr() as *mut u8,
             64,
-            NI_NAMEREQD,
+            // windows-sys 0.59 常量是 u32，flags 参数是 i32
+            NI_NAMEREQD as i32,
         )
     };
     if rc != 0 {
